Problème de sauvegarde résolu

+++ controllers/tournament_controller.py
@@ -4,7 +4,7 @@ import os
 import json
 import re
 from datetime import datetime
-from models.tournament import Tournament  
+from models.tournament import Tournament, Round  
 from config import DATA_DIR, TOURNOIS_DIR
 
 def datetime_to_string(obj):
@@ -63,9 +63,10 @@ class TournamentController:
         start_date = datetime.strptime(tournament_data.get("start_date", "01/01/2023"), "%d/%m/%Y")
         end_date = datetime.strptime(tournament_data.get("end_date", "01/01/2023"), "%d/%m/%Y")
         number_of_rounds = tournament_data.get("number_of_rounds", 4)
+        current_round = tournament_data.get("current_round", 0)
         players = self.load_players_for_tournament(tournament_data)
 
-        tournament = Tournament(tournament_id, name, location, start_date, end_date, number_of_rounds, players)
+        tournament = Tournament(tournament_id, name, location, start_date, end_date, number_of_rounds, players, current_round=current_round)
         return tournament
 
     def load_players_for_tournament(self, tournament_data):
@@ -73,8 +74,14 @@ class TournamentController:
         players = [self.player_controller.get_player_by_id(player_id) for player_id in players_ids]
         return [player for player in players if player is not None]
 
+    def load_rounds_for_tournament(self, tournament_data):
+        rounds_data = tournament_data.get("rounds", [])
+        rounds = [Round(**round_data) for round_data in rounds_data]
+        return rounds
+
     def serialize_tournament(self, tournament):
         players_ids = [player.chess_id for player in tournament.players]
+        rounds_data = [round.serialize() for round in tournament.rounds]  
         tournament_data = {
             "tournament_id": tournament.tournament_id,
             "name": tournament.name,
@@ -82,32 +89,33 @@ class TournamentController:
             "start_date": tournament.start_date.strftime("%d/%m/%Y"),
             "end_date": tournament.end_date.strftime("%d/%m/%Y"),
             "number_of_rounds": tournament.number_of_rounds,
-            "players": players_ids
+            "players": players_ids,
+            "current_round": tournament.current_round,
+            "rounds": rounds_data  
         }
         return tournament_data
 
     def get_unique_filepath(self, tournament): 
-        name = tournament.name
-        name_suffix = 1
-
-        for root, _, files in os.walk(os.path.join(DATA_DIR, TOURNOIS_DIR)):
-            for file in files:
-                if file.endswith(".json"):
-                    filepath = os.path.join(root, file)
-                    filename = os.path.splitext(file)[0]
-                    match = re.match(rf"{name}_(\d+)$", filename)
-                    if match:
-                        number = int(match.group(1))
-                        name_suffix = max(name_suffix, number + 1)
-
-        tournament_filename = f"{name}_{name_suffix}.json"
+        tournament_id = tournament.tournament_id
+        tournament_filename = f"{tournament_id}.json"
         return os.path.join(DATA_DIR, TOURNOIS_DIR, tournament_filename)
+
 
     def save_tournaments_to_file(self):
         for tournament in self.tournaments:
             filepath = self.get_unique_filepath(tournament)
-            os.makedirs(os.path.dirname(filepath), exist_ok=True)
 
-            tournament_data = self.serialize_tournament(tournament)
-            with open(filepath, "w") as file:
-                json.dump(tournament_data, file, indent=4, default=datetime_to_string)
+            if os.path.exists(filepath):
+                with open(filepath, "r") as file:
+                    existing_data = json.load(file)
+                    existing_data["rounds"] = [round.to_dict() for round in tournament.rounds]
+
+                with open(filepath, "w") as file:
+                    json.dump(existing_data, file, indent=4, default=datetime_to_string)
+            else:
+                os.makedirs(os.path.dirname(filepath), exist_ok=True)
+
+                tournament_data = self.serialize_tournament(tournament)
+                with open(filepath, "w") as file:
+                    json.dump(tournament_data, file, indent=4, default=datetime_to_string)
+
